requirements.txt
asgiref==3.8.1
attrs==23.2.0
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.7.0
